// src/api/vcci-news/models/newsletterSubscription.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * VietProDev CMS Backend API
 * Generated API documentation
 * OpenAPI spec version: 1.0.0
 */

export interface NewsletterSubscription {
  id: string;
  email: string;
  is_seen: boolean;
  /** @nullable */
  seen_at?: string | null;
  created_at: string;
  updated_at: string;
}